Fix the KVM ebtables script to stop dropping traffic to the VMM
We simply need to restrict input filtering to the corresponding VM bridge
port, as in the Xen script. Otherwise, in a standard KVM bridge setup,
traffic to the VMM physical NIC was subject to the ebtables rules and
was dropped.

This also allows the ebtables rules to do anti-spoofing between the VMs,
so correct the documentation at the beginning of the script.