Enhancement 7012: Better round-robin scheduling of multiple VMs per node.
http://bugzilla.mcs.anl.gov/globus/show_bug.cgi?id=7012

The default resource scheduler now operates with the notion of 'percentage
available' for each node in the VMM pool.  This is a percentage of the
memory previously allocated on the node and the still-available memory. This
allows the greedy and round-robin strategies to work better with pools that
have varying amounts of RAM on the VMMs.

The node selection can happen in one of two ways:

1. A "round-robin" configuration in resource-locator-ACTIVE.xml (this is the
default mode).  This looks for matching nodes (enough space to run, appropriate
network support, etc.) with the highest percentage of free space.  If there are
many equally free nodes it will pick randomly from those.  As should be clear,
this favors entirely empty nodes first.

2. A "greedy" configuration in resource-locator-ACTIVE.xml.  This looks for
matching nodes (enough space to run, appropriate network support, etc.) with
the lowest percentage of free space.  If there are many equally unfree nodes it
will pick randomly from those.