Clarify the dhcp interface selection.  See commit msg for rant.
Clarify the dhcp interface selection process so proper ebtables rules can be
chosen... on one hand it is not hard if you understand Xen networking, but
on the other hand this is still a typical impediment to installing Nimbus.
This can be done in an entirely different way so we don't force people to
understand such details (although I would personally hope a 'serious' long
term Xen VM provider will).  More on alternative solutions later."