Fix mktemp usage on Mac OS X
On OS X a template is always required.
Also, use the -t option so that the user can set TMPDIR.

Another difference in template handling between Linux and OS X will
result in strange names on OS X (e.g. tmp.XXXXXXXXXX.IgcIhfvB), but
that's OK compared to having to different code paths for this.