new DoNothingPlatform module (see commit msg)
The new DoNothingPlatform module will (not so accurately) mock a VMM, this
is mainly geared towards the new propagation-only mode for development and
testing of fast/smart propagation solutions.

The unit tests and other mock modes take advantage of the libvirt 'mock'
driver which is great for many purposes and what should be used to do
accurate VMM mocking.

What this new DoNothingPlatform module does is bypass libvirt altogether,
allowing the propagation-only mode to run on libvirt-less clusters and/or
non-privileged accounts.  Basically reducing the dependencies of propagation
research and development to the bare minimum.

See a subsequent commit for details about propagation-only mode (activated
by running src/propagate-only-mode.sh).