Change the way the mount-alter.sh lock file is managed
Move the lock used by mount-alter.sh to the Nimbus installation. The
file is already created with permissions 644. Finally, change the
flock call to open the file for reading instead of writing.

With these changes, we will be able to share the lock file between
root-owned processes and nimbus-owned processes. The write->read
part is not mandatory, it's mainly to detect a wrong installation
where the lock file would be missing (as mount-alter will not create
the file automatically anymore).