Fix fdisk output parsing
fdisk truncates the partition name when it is bigger than 79
characters. The regexp used before this commit will try to match
the full image name and will failed if the name is too long.
Fix by matching only the first 10 characters.

This bug shows up now because image names are much longer when they
are stored in Cumulus.