Proposed workaround for buggy Xen/libvirt VM state
The default Xen/libvirt on RHEL5 occasionally reports VMs as DOM_STATE_NOSTATE. The cause is not clear but anecdotally it seems to be related to multiple VMs running on a node. Previously, ws-control treated these VMs as not found which led to zombie VMs. 

This change causes info() calls to report such VMs as known but without a state.