Removed set -e from mount-alter.sh
Bash 4.x changes the behavior of "( cmd )" statements under set -e. Previously these commands would not trigger an exit but with 4.x they do. mount-alter.sh relied on the old behavior to ensure that umount is called even if the alter fails. Other commands in the script are checked for success, so we are safe in removing set -e.

Closes GH-18.