Fixed problem backing out VMM memory allocation
Related to #46 but may not solve it entirely. The `in_use` check just compares memory available to max memory and there could be other sitations which cause a leak of VMM memory use.

It happens when an exception is raised during NIC binding (such as a NIC requesting a nonexistent network). No attempt was made to back out the scheduled reservation. Solved by doing NIC binding before scheduling but I'm not totally clear on any side effects this might have.