adding logic to retry the ssh command which accesses that database.
the sqlite db can through errors on db contention.  because this command happens when all the file is completed it is a likely place for contention under some circumstances.  When transfering and image to multiple destinations on a single VMM it is likely that they will all complete at near the same time.  This patch adds logic to retry the command with backoff

another patch to AsyncNotification.py will limit the size of the error text.