Update cumulus dependencies to pyOpenSSL 0.13
The latest Ubuntu releases do not include SSLv2 support in the OpenSSL
libraries, because of security flaws in the protocol. This makes cumulus
fail to install on a recent Ubuntu, because pyOpenSSL 0.10 tries to use
the SSLv2 symbols. This commit updates the dependency to pyOpenSSL 0.13
which removes SSLv2 support if the underlying OpenSSL library does not
provide it.