Install pyOpenSSL 0.10 when OLD_OPENSSL_VERSION is defined
We recently updated pyOpenSSL to version 0.13 in order to fix a build
problem on recent Ubuntu releases. However, 0.13 does not compile on
Centos 5 because is requires OpenSSL 0.9.8f or later.

If the OLD_OPENSSL_VERSION environment variable is defined, use a
different requirements.txt file which installs pyOpenSSL 0.10.

The cumulus-deps tarball was updated to include both pyOpenSSL packages.