Add a partition label to each blankspace partition
Labels are extracted from the partition file name, and are named
incrementally: blankpartition0, blankpartition1, etc.

A label on the blankspace partition will allow mounting via label
instead of device. This makes it easier in fstab where the device of the
blankspace partition may not be known a priori.

Thanks to Michael Paterson for the feature request and initial patch.

Closes #88.