Add support for booting KVM machines from qcow2 images
Detect qcow2 images by reading the image file header.  It contains the
"QFI" string and the qcow version number.  The libvirt template is
configured to use the qcow2 driver when a qcow2 image is detected.

The root partition is altered by attaching it to a host device using
qemu-nbd.  Since qemu-nbd can be named differently depending on the
Linux distribution, it is configurable by mount.conf.

Closes #105.