Add copy-on-write support
When this feature is enabled, a copy-on-write volume is created to store
modifications to the virtual machine image.

If the image cache is also enabled, images are directly hard linked from
the image cache into the virtual machine secureimages directory.  These
images stay unmodified because all writes are stored in the
copy-on-write file.

Since this feature is experimental and needs careful management of disk
space, it is disabled by default.

Closes #106.