Try to stop all instances when running nimbus-admin --shutdown --all
Corrupted instances have shutdown capabilities disabled.  When
--shutdown --all is called, it would stop at the first Corrupted
instance with the following message:

    RemoteException occurred in server thread; nested exception is:
    java.rmi.RemoteException: Shutdown is currently disabled

Catch the exception and continue to the next instance.  A warning is
printing in the logs and a message is returned to the user.

Also add a --force argument to nimbus-admin to activate the trash
operation for instances that didn't shut down properly.